remove bad folder and move Calc.java to src
