--- COMP354_Calculator-master/README.md
@@ -1,2 +1,0 @@
-# COMP354_Calculator
-ETERNITY: A calculator project for COMP 354.
