--- COMP354_Calculator-master/nbproject/private/configs/Run_in_Browser.properties
@@ -1,2 +1,0 @@
-# Do not modify this property in this configuration. It can be re-generated.
-javafx.run.as=embedded
