--- COMP354_Calculator-master/nbproject/private/private.properties
@@ -1,4 +1,0 @@
-auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
-# No need to modify this property unless customizing JavaFX Ant task infrastructure
-endorsed.javafx.ant.classpath=.
-user.properties.file=/home/b/b_hua/.netbeans/8.0.2/build.properties
